test(scripts): extract coverage-gaps arg parser and cover it
scripts/audit-coverage-gaps.mjs parsed its --json/--min-demand/--per-category
flags with a local parseArgs that had no tests. Hoist it into the existing
scripts/lib/coverage-gaps.mjs (which the script already imports for
findCoverageGaps) as parseCoverageGapsArgs and use it from the script.

- scripts/lib/coverage-gaps.mjs: add parseCoverageGapsArgs.
- scripts/audit-coverage-gaps.mjs: import the shared parser; drop the local copy.
- tests/coverage-gaps.test.ts: add cases for the flag defaults, --json plus the
  numeric flags, the non-numeric / zero fallback to 12, and ignored tokens. The
  new function's lines and branches are fully covered.

No behavior change; the CLI arguments parse identically.
